requirements.txt
allure-pytest==2.6.2
allure-python-commons==2.6.2
atomicwrites==1.3.0
attrs==19.3.0
CacheControl==0.12.5
cachy==0.2.0
certifi==2019.9.11
chardet==3.0.4
cleo==0.6.8
colorama==0.4.1
colorlog==4.0.2
et-xmlfile==1.0.1
filetype==1.0.5
har2case==0.3.1
html5lib==1.0.1
httprunner==2.3.2
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
Jinja2==2.10.3
jsonpath==0.82
jsonschema==3.1.1
lockfile==0.12.2
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
msgpack==0.6.2
openpyxl==3.0.1
packaging==19.2
pastel==0.1.1
pkginfo==1.5.0.1
pluggy==0.13.0
poetry==0.12.17
py==1.8.0
pylev==1.3.0
pyparsing==2.4.5
pyrsistent==0.14.11
pytest==5.2.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
shellingham==1.3.1
six==1.13.0
tomlkit==0.5.8
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==0.6.0
